Refactored ippc_server folder
